Attempt to preserve upload order when uploading multiple files.
File uploads will get a semi-random ID attached to them, with subsequent
files referencing that ID as "parent", when adding files to the juggler,
and the parent ID isn't in the list yet, the file will wait for the
parent to be added first, for up to 30 seconds. If the 30 seconds pass,
the file will not be added.